# feat(batch): add `ordered` property for seq scan