--- a/Cargo.toml
+++ b/Cargo.toml
@@ -101,6 +101,8 @@ opt-level = 3
 opt-level = 3
 [profile.ci-dev.package."indextree"]
 opt-level = 3
+[profile.ci-dev.package."task_stats_alloc"]
+opt-level = 3
 
 # The profile used for deterministic simulation tests in CI.
 # The simulator can only run single-threaded, so optimization is required to make the running time