--- a/src/compute/tests/integration_tests.rs
+++ b/src/compute/tests/integration_tests.rs
@@ -220,6 +220,7 @@ async fn test_table_materialize() -> StreamResult<()> {
     let scan = Box::new(RowSeqScanExecutor::new(
         table.clone(),
         vec![ScanRange::full()],
+        true,
         u64::MAX,
         1024,
         "RowSeqExecutor2".to_string(),
@@ -282,6 +283,7 @@ async fn test_table_materialize() -> StreamResult<()> {
     let scan = Box::new(RowSeqScanExecutor::new(
         table.clone(),
         vec![ScanRange::full()],
+        true,
         u64::MAX,
         1024,
         "RowSeqScanExecutor2".to_string(),
@@ -354,6 +356,7 @@ async fn test_table_materialize() -> StreamResult<()> {
     let scan = Box::new(RowSeqScanExecutor::new(
         table,
         vec![ScanRange::full()],
+        true,
         u64::MAX,
         1024,
         "RowSeqScanExecutor2".to_string(),
@@ -421,6 +424,7 @@ async fn test_row_seq_scan() -> Result<()> {
     let executor = Box::new(RowSeqScanExecutor::new(
         table,
         vec![ScanRange::full()],
+        true,
         u64::MAX,
         1,
         "RowSeqScanExecutor2".to_string(),