I hereby agree to the terms of the [Singularity Data, Inc. Contributor License Agreement](https://gist.github.com/skyzh/0663682a70b0edde7ae991492f2314cb#file-s9y_cla).

## What's changed and what's your intention?

This PR adds the `ordered` property for seq scan. If the optimizer does not require the order of seq scan, we can utilize the optimization in #6927.

Rough test: this PR reduces the latency of selecting an empty table from 30ms to 15ms under `dev` and 4.5ms to 2ms under `release`.

## Checklist

- [x] I have written necessary rustdoc comments
- [x] I have added necessary unit tests and integration tests
- [x] All checks passed in `./risedev check` (or alias, `./risedev c`)

## Refer to a related PR or issue link (optional)
- #6927 
- #5594